backend/requirements.txt
Flask
Flask-Cors
gunicorn
celery
redis
psycopg2-binary
Flask-SQLAlchemy
reportlab
opencv-python-headless
pytesseract
numpy
Pillow
requests
